Remove deprecated secrets file
The use of Rails.application.secrets is deprecated and will be removed
from Rails 7.2.

There is no longer a need to use a secrets.yml file for setting
secret_key_base. Rails automatically generates a secret in development
and test environments (which is stored in a tmp/local_secret.txt file)
and automatically reads SECRET_KEY_BASE from an env var [1].

[1]: https://api.rubyonrails.org/classes/Rails/Application.html#method-i-secret_key_base