Include description in the chunked_content_index
This was a field I missed when I implemented this the first time and is
intended to be used as part of the content that is indexed and searched.

Unlike other key fields description is typically an optional field in
the Publishing API and thus we need to cater for it being nil.