Document Heroku approach to populating OpenSearch
This documents the rather nuanced process that I've just been doing on
my own laptop. There's lots to improve here but it seems better than it
only existing in my head.

I've left development in a TODO state as I'll plan to add that in the
next PR in this story.