Move ApplicationController logic into BaseController
We're going to add an admin interface to the application. As the controllers
from both interfaces will inherit from the application controller, each
interface requires it's own base controller for interface specific logic.

This commit moves the logic from the application controller for the
"require_chat_risks_understood" method to the new
base controller.