Add basic layout for admin interface
I've not added flashes or the error summary component into the layout
because as far as i can tell they're not required in the next couple of
cards.

My thinking is we can add it later if they're required. But happy to
add it now if people disagree.

I've was getting some weird test failures on a follow up PR as it was
looking for favicon-test.png.

Looking at Content Publisher i was able to see the environment being
passed into the layout_for_admin component was
GovukPublishingComponents::AppHelpers::Environment.current_acceptance_environment

This passes "development" when the env is "test". I've updated the
application layout to use this too.