Use chunked_content data seeds generated by ChatGPT
This introduces a set of seed files that are generated by ChatGPT
(approach documented in the included README file).

These have been added as a means to store data that can be used to
populate the chunked_content OpenSearch index in a development
environment. We want to have a means for this because having some search
records is essential for usage and the process to populate with actual
GOV.UK data is arduous [1].

ChatGPT has been used to generate a set of seed files based on ChatGPTs
idea of GOV.UK content. This data may prove a bit contrived, but
hopefully it's a reasonable starting point for dev usage.

I skipped generating either content_id or locale, as ChatGPT doesn't do
a fantastic job at generating valid UUID values and locale is always
"en".

This is an alternative approach to that raised in: https://github.com/alphagov/govuk-chat/pull/116
which used a series of gzipped files of GOV.UK content.

This approach is preferable because:

- The data is human editable which makes it easy to read and changeable
  should the search schema change
- Much less data, doesn't add 200MB to the repo, but then less search
  results and much faster to use
- Isn't real GOV.UK content so less representative of actual GOV.UK chat
  experience, but then doesn't carry the risks of real, yet out-of-date,
  data

[1]: https://github.com/alphagov/govuk-chat/blob/main/docs/message-queue-consumption.md#consuming-queues-in-a-development-environment