Add tests for chunked_content seeds
Since we rely on ChatGPT to generate this data I thought it would be
prudent to write some tests to assert this data is as we expect, with
the expectation that this will save someone time in having to check all
the data.

I did have to do some unusual things to achieve this though so this
could be a divisive approach. It is a bit unusual to write tests for the
db folder of Rails and I'm using rspec-mocks as a pseudo validation
tool as I wanted to have something lightweight with the tools I had to
hand.

Finally the use of `specify` in the spec may be surprising. This is done
because the tests don't read write using `it` as per rspec styleguide
[1].

[1]: https://rspec.rubystyle.guide/#it-and-specify