Add and test rake task to load chunked_content data seeds
This assembles a rake task that can take the data from the
chunked_content seeds files and update the OpenSearch index with them.

The expected use case for this is developers who are populating their
local system with data so they have a usable chat that can find data to
some questions.

This rake task is potentially a bit fragile in it's coupling to
application code. It makes use of a few classes that weren't designed
for this use case and would be vulnerable to break should the needs of
those classes change. I did this as I thought, despite this
disadvantage, that it was still advantageous compared to having to
duplicate a bunch of logic.

The rake task runs the create_chunked_content_index as a pre-requisite
as, if a user is missing the index, we don't want OpenSearch to just
create it's own embeddings for the index.

As we don't want to have conflicts on content items with the same
base_paths (and have application logic to stop it which is bypassed here
[1]) we first do a clear out of anything with the same base_path as any
of the data seeds.

[1]: https://github.com/alphagov/govuk-chat/blob/32f4609f4efbbde3dc7bfb9489733fece20ed850/lib/message_queue/message_processor.rb#L15-L26