Add documentation on populating search in a dev
This documents the usage of a rake task to populate the search index
based on the seed data and links to a more advanced option for other use
cases.

Ideally in future it'll be easy enough to clone the data from
integration to have the full search index locally, should that be
useful.