Introduce BodyContentArrayParser and move 'answer' schema to use it
This is a simple schema - no include/exclude list so a good candidate to move
over to the new simple parser.

The intent is to move all the schemata that have an array body with
text/govspeak and text/html to use this class and then remove this logic
from the BodyContentParser.