Concurrently pull data from the message_queue
This makes use of options that were introduced in
govuk_message_queue_consumer v5.0.0 [1] and uses the same environment
variable names as Search API V2 [2]. These options allow the worker
processes to consume messages concurrently.

I've been more bullish than Search API v2 and set the defaults of this to
be 10 threads. I did this because it can be very slow to read a lot of
Publishing API data (as per [3]) and I tend to end up running the
process multiple times on my machine so it runs faster.

As per Chris' suggestion [4] I've also set the prefetch number to match
the number of threads.

I expect these env vars to be somewhat tuned for running on GOV.UK
infra.

[1]: https://github.com/alphagov/govuk_message_queue_consumer/blob/main/CHANGELOG.md#500
[2]: https://github.com/alphagov/search-api-v2/blob/3dab422e24bb16b1c38131d40280f7a0be13e174/lib/tasks/document_sync_worker.rake#L22-L23
[3]: https://github.com/alphagov/govuk-chat/blob/0e6e8ea9b2601740a0cf8d20a02420e589701166/docs/message-queue-consumption.md#consuming-queues-in-a-development-environment
[4]: https://github.com/alphagov/govuk_message_queue_consumer/pull/109#issuecomment-2082437782