Fix incorrect Feature guard clause
This guard clause had an error where it was checking Current.user rather
than just the user argument. It turned out the guard clause itself
wasn't necessary as Flipper responds the same with a user of nil.

As this had slipped through I thought it was better to update the tests
to integrate with Flipper directly rather than stubbing it, so that we
could be clearer that this works as intended.

It does however leave a rather sad test to code ratio. Sorry.