Add a flipper_id method to a user
This allows the email address to be set as a flipper id making it quite
easy to set users to have access to a feature by the Flipper UI.

I didn't think we'd need to have the more complex structure that Flipper
uses of "User; <id>" as I don't expect we'll use flipper for anything
more than user models.

As email is a field that can be null (though in practice seems unlikely)
it falls back to id.

I'm also unconcerned about whether a user has multiple accounts with the
same email.