Use GDS SSO user instead of AnonymousUser
Now that we have GDS SSO installed we can use that instead of the
AnonymousUser system we have set-up.

As part of this I also replaced the stubbing of Feature with just making
a call to Flipper. This seemed more appropriate for a system and was
very low on the amount of code needed.

There was a bunch of discussion before hand about whether we should have
a null object pattern for an anonymous user. I explored this idea in [1]
but felt that we don't make enough use of user now to warrant it, so
parked it as an idea to explore when we do more with a user.

[1]: https://github.com/alphagov/govuk-chat/commit/5921e86be5a66b51c02112bcd40c91b4bb1b6108