Include specific request test for feature flags
In my previous commit I simplified the system test to no longer utilise
feature flags at a user level. I felt that it was appropriate to have
some integration tests to flag the whole process for a feature flag
based on an individual user, it seemed appropriate to this at a request
level.

One thing that was very challenging was working out how to set-up the
mechanism to stub whether a user was logged in or not. It was hard to
find a precedent in GOV.UK apps of how to do this as the approaches I
found either: only worked if login was mandated [1] or worked only on
lower level controller tests [2].

Therefore I came up with an approach that made it possible to stub the
controller. It did mean committing the sin of using
`allow_any_instance_of` but I'm not sure there is a nicer way to achieve
it. It'd be slightly nicer to stub the env["warden"] call instead of the
warden method, but that seemed trickier to mock as it'd mean stubbing a
hash on a request object.

[1]: https://github.com/alphagov/content-publisher/blob/c15c31a5caefc00aff0f41f483c7a9baea5bffaa/spec/support/authentication_helper.rb
[2]: https://github.com/alphagov/local-links-manager/blob/d0768305ea5f87449dc1804ad6bf3cb74545c5b4/spec/support/authentication.rb