Add code_snippet component
This is a super simple component which takes some content and renders
it within inside a code tag within a pre tag.

I've pulled the styling from Content Publisher here
https://github.com/alphagov/content-publisher/blob/c15c31a5caefc00aff0f41f483c7a9baea5bffaa/app/assets/stylesheets/components/_markdown-guidance.scss#L52-L57

It really only needs some content so i haven't bothered adding in a bunch
of stuff like id, data attributes etc. as they're unlikely to be needed.

We can always add them in later if we want to.