Add pry-byebug
This adds the ability to navigate while in a debugging session as per
byebug.

I'm of the view that it'd probably be better to ditch pry-rails and
pry-byebug for the debug gem but I know some team members struggled with
debug last time so will leave that as something to consider later.