More robust seed data
This changes the User seed to only have limited data in the
find_or_create_by so changes in this hash don't create extra records.

The data that is liable to change (the random uuid and permissions) is
now set to always be updated so user's tests don't slip out of sync with
this.