Ignore schema_type of substitute from search ingestion
Publishing API recently introduced a new feature that broadcasts when a
piece of content has been unpublished as a substitute [1]. A substitute
unpublishing occurs when a piece of content has been replaced, for
example if a document is replaced with a redirect, or a different
document.

This feature could cause us problems for data integrity as the Publishing
API will emit two events for a particular base_path and we should only
index one of these. As these events are only the result of another event
occurring it should be of no negative consequence to ignore them.

[1]: https://github.com/alphagov/publishing-api/pull/2664