Update format_answer_status_as_tag produce descriptive output
This commit updates the format_answer_status_as_tag helper method to
provide additional context to the user on the status of the answer.

When called from the status column of the table it sets the title
attribute of the tag to the status of the answer and the description
if one is returned from the case statement.

On the question show page, it returns the same tag wrapped in a span with
a visible description of the status if one returned from the case statement.