Add component-guide stylesheet which loads admin and application stylesheets
We've run into a bit of an issue with our local component guide. Currently
it defaults to using application.css. This causes an issue as some of our
components are interface specific, so are only loaded in application.css or
admin.css.

This commit adds a new component-guide.scss file which imports both files
I could just manually import the components we've added, but then we
would have to keep it in sync with the application.css and admin.css files.

The load time will be slightly slower, but since we're only using in local
dev I don't think it's a big issue.