Add scope with tags and context to Sentry for message queue consumer
We rescue from RuntimeErrors and exceptions that are raised with our OpenSearch and
OpenAI integrations while the message queue consumer is processing
a message.

This commit adds context and tags to Sentry so that we've got much fuller
picture of what's happening when we get an error.

This will help with debugging and understanding edge cases that might occur.

We should use the messages payload to set the following for the context

- content_id
- locale
- base_path
- document_type
- payload_version

And add a few tags for ease of filtering

- content_id
- base_path