Add new basic routes and views for onboarding flow
- updates the existing onboarding actions to limitation endpoints
and adds basic view
- sets session[:onboarding] to "privacy" when the user hits the
confirm_onboarding_limitations endpoint
- adds onboarding privacy endpoints
- ets session[:onboarding] to "conversation" when the user hits the
confirm_onboarding_privacy endpoint
- updates the tests in the chat request spec to reflect the new
endpoints
- renames the BaseController#require_chat_risks_understood to
BaseController#require_onboarding_completed
- updates the logic of the above method to use the new session[:onboarding]
to determine whether onboarding needs to be completed.

This does not include any of the redirects to ensure the user is on the
correct page based on the state of the cookie and session. This will
come later down the line.