Add before_action redirects onboarding endpoints
We're adding in some redirects to the onboarding endpoints to ensure that
the user is redirected to the appropriate part of the onbarding flow or
to their conversation if they've already completed the onboarding.

These should behave in the following way for the limitation endpoints:

```
if conversation_id is present on the cookie
  redirect_to conversation show page if
elsif cookie is not present && session[:onboarding] == "conversation"
  redirect to new conversation page if
elsif session[:onboarding] == "privacy"
  redirect to privacy page
end
````

and the following for the privacy endpoints:

```
if conversation_id is present on the cookie
  redirect_to conversation show page if
elsif cookie is not present && session[:onboarding] == "conversation"
  redirect to new conversation page if
elsif session[:onboarding] != "privacy" || "conversation"
  redirect to limitations page
end
```

These will able to be simplified once we have merged the new and show page
because the first two redirects will be merged and always redirect to
the show page

I've pulled out as much of the logic as possible into shared example.
Some of these will be reused be used for the conversation request spec
when we migrate the conversations controller to using the cookie approach.