fix brakeman error
Brakeman seems to have got itself confused and thinks we are trying to render
the :id parameter onto the page, failing with this

Confidence: Weak
Category: Cross-Site Scripting
Check: CrossSiteScripting
Message: Unescaped parameter value
Code: Search::ChunkedContentRepository.new.chunk(params[:id]).html_content
File: app/views/admin/chunks/show.html.erb
Line: 10

Brakeman is wrong - even if someone could hack an injection attack into the :id parameter, it
wouldn't match a chunk id and would just result in a 404, it's never going to
get rendered onto the page so is not a security issue.

Wrapping it in a method seems to make brakeman happy - I think this is probably
a bug in Brakeman as this is semantically identical.