Add logic to render more information on limitations page
This adds in some conditional logic to handle rendering additional
information on the page when the user click the tell me more button.

It uses a query string to determine whether to render the additional
messages to the user and whether the render the tell me more button or
a link to GOV.UK.

It also sets session[:more_information] to true in the controller when
the user arrives having clicked on the tell me more button.