Scope govspeak component override to message component
The govspeak override was applied without a component scope so it's
effects were global. This changes moves this change to be within a class
of this components scope so that this override is isolated to just this
component.

I also updated the component guide doc for this to have this scenario
covered so it could be seen visually.