Configure RSpec for JS system tests
This allows the application to run JS system test with the driver
configured by govuk_test.

I've set this up after hitting issues with JS tests in a wider PR [1]
and this seemed an opportunity to break that work up.

There are no tests that use this yet, so there isn't a guarantee this
will continue working.

[1]: https://github.com/alphagov/govuk-chat/pull/181