Simplify the conversation_input component
This removes options many of the options provided to this component on
the basis that this component is only going to be used in one specific
circumstance and thus can be defined statically.

I've switched the ids to just have a single one that may be generated
and then subsequent ones suffix that.

This also fixes accessibility warnings flagged in the component guide by
correcting values for the id and it changes the RSpec tests to use the
more idiomatic syntax.