conversation-input -> conversation-form
This modifies the conversation-input component to contain a form instead
of this being wrapped outside it. It then modifies the component name to
reflect that is now a form.

This has been done to support introducing JavaScript into the process of
submitting questions as explored in [1] where it would be simpler for
this to be a single component.

This is a precursor to work that adds JS events into this form.

Apologies to James who initially implemented this component as a form
and I persuaded him to roll back on that.

[1] https://github.com/alphagov/govuk-chat/pull/181