Simplify initialising OpenSearch
This refactors the code used to initialise OpenSearch to be simpler. The
motivation for this change was that, by using Rails.env.production?, we
had code that did not have test coverage.

It isn't a concern that in all environments there will be user and
password values that are nil, we already have this scenario in
production in Heroku where user and password is done via the URL
directly. I'm not too sure why we didn't do that in AWS - maybe special
characters in the password?