Add parent document type on search ingestion
As part of the re-reranking that the DS's are testing, they use the parent
document type of html publications during the re-ranking process so that
they are correctly weighted.

This adds the parent document type to the search ingestion process so that
we can it when we implement the re-ranking.

I've updated the content_item_chunk_factory & notification_content_item_factory
to handle being passed in a parent_document_type.

When present it overwrites the `content_item["expanded_links"]["parent"]
which a hash containing the parent document type and the required fields.