Format plain content when rendered
Prior to this all this content was shown on a single line, by using
simple format this content is converted to a paragraph with line breaks
on each new line.

I put this inside the govspeak component so that it would render a
little nicer, as without it there was quite a big top margin on the
paragraph.