Tweak chunk rendering for ease of use
From using the search I felt I could improve the rendering of chunks
further by moving the URL and chunk number up in prominence and moving
the Search content down to be below the HTML content's prominence (as
the latter feels more important and the former can be repetitive).

I also felt that a chunk number of 0 didn't tell us a lot so thought it
better to show a x of y number.

I imagine the next logical step is to show a view where you can see all
the chunks for a base_path, but I thought that can be done when there is
more clarity in this being a user need.