Add JS classes for component-form fields
These classes are going to be used as part of JavaScript for the
conversation-form component. They have been created to provide selectors
that the module for the component-form component can use to look up
these items based on my understanding that GOV.UK convention is to use a
js- class over an app-c- one.