Add JS for conversation-form events
This adds some basic JS to the conversation-form component to handle a
few events that are anticipated to be needed (based on [1])

The JS has been written using modern JavaScript syntax with the
expectation that we will be dropping IE11 support and thus can make use
of it.

This commit hooks the JS up into the component gem as well.

[1]: https://github.com/alphagov/govuk-chat/pull/181