Add Jasmine JS tests and related config
This adds in the configuration needed to run Jasmine JS tests (Jasmine
was chosen as per GOV.UK Rails conventions [1]) and some basic tests
which test the functionality of conversation-form.

Like the implementation code, I have used modern JS syntax for writing
the tests with the expectation they won't be needed to be executed in
legacy browsers.

The tests I've added are all pretty simple and just serve as a quick
example of setting up Jasmine.

[1]: https://docs.publishing.service.gov.uk/manual/conventions-for-rails-applications.html#testing-javascript-code