Use standard rather than standardx
We no longer have a need to use standardx now that we're not doing any
eslint configuration so can go back to regular standard as per GDS Way
conventions [1].

Unfortunately the shared workflow for JS linting isn't flexible [2] so
we have to define a GitHub action manually. Ideally the shared workflow
would give us the freedom to specify the package.json task so it could
be used for pretty much any node task.

[1]: https://gds-way.digital.cabinet-office.gov.uk/manuals/programming-languages/js.html#linting