Validate question presence client side
This replaces the use of the required attribute on the conversation-form
with client side validation.

It uses a data attribute to pass in the validation error message so that
this can be managed server side (we could use a locale file to have this
message defined in only one place).

We no longer use the gem component for errors as we want to
manipulate the contents directly via JS. On review it was then suggested
we use a unordered list for improved semantics [1].

I expect the aria attributes may need additional work on this. As the
error wrapper is always rendered I removed the conditional on setting
aria-described-by and instead use the hidden attribute to hide/show the
element as needed. I wasn't sure if this would be ok for a screen-reader
and did expect we might need an aria-live attribute.

I did consider modifying aria-describedby dynamically, but this is a bit
of a pain as the attribute is not a DOMTokenList and would need some
string manipulation.

Finally in configuring this I realised that it is a bit dangerous for
another module to set a submit event listener on this form as it might
not be deterministic as to which event fires first. I imagine solutions
to this are:

- having another module listen on an element up the tree (to catch it
  when bubbled)
- have the form fire a custom event which can be listened to.

[1]: https://github.com/alphagov/govuk-chat/pull/247#discussion_r1617041948