Remove unnecessary notification_content_item validation step
We've learnt that the validation process for GovukSchemas are quite time
expensive and are making up a relatively high proportion of the time to
run the test suite.

It turns out that we didn't actually have to do this step ourselves as
this is functionality that is built into GovukSchemas [1]. This change
modifies our code to instead make use of this functionality and remove
our own validation.

This code uses a proc so that this can be executed as a block for
implicit validation when using the payload method, if not ensuring the
item is valid we don't use the payload method and don't use that
implicit validation.

I don't think this really produces much of a net-reduction in validation
runs however it does speed up the test suite for me locally. Using
govuk-docker with my MBP I get ~40s with this change and ~55s before it.

There is further changes that could be made to GovukSchemas to make
things faster. The payload code [1] is configured to validate a schema
twice, we could change this so:

1. The validations are controlled with arguments
   (validate_base_and_customised: false perhaps)
2. Checking if an item is actually customised before validating it again

[1]: https://github.com/alphagov/govuk_schemas/blob/0cb2983c38770075ed80a18f6973cd1bad70a923/lib/govuk_schemas/random_example.rb#L83-L97