Add a conversation layout
This layout sets a variable that modifies a number of the base classes
of the application layout in order to adjust the layout to be one with
overflow-y disabled at the HTML element.

It replaces the global overrides we had with the use of classes in the
conversation-layout object.

This has been done so that pages that don't use the conversation layout
can be unaffected by unconventional changes - such as scrolling no
longer done on the HTML element and so that we don't need to undo global
styling whenever we work on a page that isn't part of the conversation
layout.