Move conversation layout structure to layout file
This moves layout code from the conversations show template to the conversation
layout. It then makes use of this code for the onboarding views which
have the same structure.

Doing this allows us to cut out repetition in the individual views and
have these similar views rendered consistently. This is expected to be
useful when it comes to setting up the JS experience for these pages
where we will need a consistent structure.