Block legacy browsers from running JS
This adds JavaScript for the app to be run within <script type=module>
with the application.js and the inline js-enabled block using it.

We have an application-legacy.js which is served to all browsers with a
regular script tag. This is needed because the scripts from
govuk_publishing_components do not work under <script type=module>
because they make use of `this` as a global variable [1].

We therefore use a technique that we've copied from Whitehall [2] that
uses a trick with the useCapture argument to take priority over the
existing eventListener for DOMContentLoaded so that this handler can run
first and prevent further propagation.

[1]: https://github.com/alphagov/govuk-rfcs/pull/171#issuecomment-2114348272
[2]: https://github.com/alphagov/whitehall/pull/7352