Add questions filter to admin pages
This adds the ability to filter by status on the questions index and
show conversation pages in the admin interface.

It adds a questions filter class to handle the filtering

Currently this only has the abiity to filter by status, but in follow up
PRs i'll add in the ability to filter on date and the ability to search
for a question.

I've also added a concern to handle the filter params since these are
being used in multiple controllers.

There's probably an argument for moving the conversations_controller#show
endpoint into the questions_controller since it's the only action in
the conversations_controller.

Potentially named something like QuestionsController#show_conversation.

One thing that isn't working in this commit is the pagination links once
the status is filtered.

This will be fixed in the next commit.