Set an override for govuk-template__body
This resolves an issue that was seen in Firefox where a page with a
short amount of content reveals the page background colour in the
footer.

We've resolved this by setting the minimum height of the body element to
be 100% of the viewport size. We felt that this offered a more
consistent GOV.UK experience than changing the background colour of the
html element to white.