Create a vertically-centred-layout CSS object
This object is used so that any contents in the main element will be
aligned vertically centred. This is to maintain the behaviour for the
homepage where the introduction component is rendered with vertical
centre alignment.

I spent far too long wondering whether the object should be named in
British (centre) or American English (center). I felt that as a UK
government civil servant, I should go British. I do appreciate the risk
of confusion here.