Allow errors to be injected as part of question rejection
This is code to allow server-side generated errors to be rendered on the
client side.

The expectation is that when we receive a response from the server that
a question has been rejected we fire the question-rejected event and
include a detail of an array of errorMessages.

I struggled with trying to catch the error as i found doing techniques
like:

```
expect(() => form.dispatchEvent(new Event())).toThrow()
```

or

```
try {
  form.dispatchEvent(new Event()))
} catch (e) {
  console.log(e)
}
```

didn't catch the error as the error was thrown outside of the scope of
the executing code. I couldn't find a solution to this so had to resort
to calling the method directly on the module.