Reduce number of parser tests
This removes loops of tests that test the configuration of these parser
classes rather than logic defined within the code. This change
simplifies the cases where we loop through every input that could
produce an asserted output to just checking a single one.

The motivation for this is that we recently learnt that the random
content item generator is quite time expensive so it seems appropriate
to remove any cases where we've got tests that don't clearly test
application logic.

Performance impact:

Before:

```
➜  govuk-chat git:(audit-schema-usage) rspec spec/lib/chunking/content_item_parsing/

Randomized with seed 30844
..............................................................................................................

Finished in 16.29 seconds (files took 3.32 seconds to load)
110 examples, 0 failures

Randomized with seed 30844
```

After:

```
➜  govuk-chat git:(audit-schema-usage) ✗ rspec spec/lib/chunking/content_item_parsing/

Randomized with seed 31254
..........................................................................

Finished in 8.38 seconds (files took 3.99 seconds to load)
74 examples, 0 failures

Randomized with seed 31254
```