Remove document_type handling from parser shared example
It turns out these document types weren't being applied as the argument
to create the content item didn't have it as a parameter. If this had
made a difference we'd have seen flaky tests, however as any document
type handling is distinct from the ability to parse this information
doesn't matter and we can therefore simplify the tests.