Default missing args to nil in Search::ChunkedContentRepository::Result
This fixes a production issue where, since the introduction of the
parent_document_type field, any initialisation of
Search:ChunkedContentRepository::Result has raised an error of:

```
missing keyword: :parent_document_type (ArgumentError)
```

This error occurred because we had a new field to the search index but
this data is not retrospectively set on existing index entries. As the
Data object required this field any code that initialised this object
raised an error.

This change sets this object to initialise all attributes with a value
of nil if they are missing which prevents this issue from occurring. I'm
expecting we will want this for any future changes to the search index.
It's a bit of a blunt object, as we could actually want an exception if
some fields aren't set - but given there is no validation of data going
into this object it seems of low consequence.