Replace repetition of error messages
This allows both the client and server side error messages for
Form::CreateQuestion to use the same strings and thus only be defined in
one place.

I chose the approach of adding constants as a way to share these
messages as I felt that this was the approach that was most consistent
with this codebase.

There were a couple of rejected options:

1) Use I18n, Rails provides a nice way to define ActiveModel error
   messages in the locale.yml file. We could look write them once there
   and look them in view context, and they'd be automatically loaded for
   ActiveModel. I rejected this as we make no use of Rails I18n
   currently and it would likely cause confusion as some error messages
   would be defined there and not others.
2) Try and look up the messages directly from the model - there isn't
   really an API for this so it feels fragile to try do it (e.g.
   Form::CreateQuestion.validators then try traverse until you get an
   object with an options of message)

[1]: https://guides.rubyonrails.org/i18n.html#translations-for-active-record-models