Move answer status label and description into config
We're starting to use some logic the various answer statuses throughout
various parts of the admin interface. This commit moves the label,
description, and color for each answer status into a config file so they
can be reused.

I've also added a label_and_description which is passed to each option
in the status select in the fileter partial.