Add search to QuestionsFilter
This adds the ability for the questions fitler to take a search input.
It takes this and searches against:

- the question messsage
- the answer message
- the rephrased question

There's a weird bit of behaviour here that happens. I've explicity used
a left joins and set the table alias to answer because otherwise you can
get some weird bugs, when you use active record to search on a has one
association.

Active record assumes that has one association tables have been aliased
to singular, i.e. the answers table has become answer.

This can cause the filters to blow up when different combinations of filters
are used, as some expect singular and some expect plural.

Ensuring it's set to singular and updating our custom sql to use the alias
feels like the simplest way to circumvent this.