Update ErrorsHelper#errors_for to take mapping of attrs to anchors
We need to pass errors for the QuestionsFilter form to the error summary
component for multiple attrs. This updates the errors_for helper to take
a hash with attrs as the key and anchors as the value.

These can be used to set the href of the error link.

It updates the conversation show page to use the new implementation.