Call validate as last step of initialisation
As we're not modifing any of the attrs post initialistion, calling
validate at the end of init ensures that the errors are set on the
filter model.

These will be used in the error summary component and date inputs in
the view. It feels cleaner to do it this way than call validate in the
controller.