JS to handle a question submission
This configures the basic handling of a question via JS and redirects a
user to an answer by updating the window.location.href.

In order to set the data module for this, we have moved the
app-c-conversation-layout__wrapper class into the conversation layout
before adding a data attribute. The move of this was done on the
expectation that we may have to modify this element more and that it'd
be better done without being two view files removed and in the global
layout.

This sets up a number of basic tests for the module, with some caveats
for testability:

1) We couldn't establish a way to test the asynchronous handling of an
   of asynchronous operations caused by an event. After much googling
   and tweaking we found the only reliable way to test it was to test
   the asynchronous event handler directly as we could await that.
2) We don't know of a way that we can mock the setting of
   window.location.href for tests. Therefore we've added an extra method
   that can be stubbed so that we can assert.