Add QuestionsFilter#toggleable_sort_params
This methods handles constructing the toggleable sort params used in
the message and created_at columns of the tables that render questions.

It takes an argument which is the default sort value for that column.
If the current value of the sort attribute doesn't match that value,
then it constucts the link with the default sort value as the sort param.

If it matches the value, then it constructs the link with either:

- desc if the current sort value is asc
- asc if the current sort value is desc

One thing to note with the tests here is that the regexes that matches
the question messages are in the correct order:

```
/Hello world.*World hello/m
/World hello.*Hello world/m
```

require the /m flag to match correctly as the values are split over
multiple lines/rows in the table.