Polling for an answer via JS
This sets up a polling mechanism when there is an answer pending for a
conversation. This polling will check every 500ms for the availability
of an answer and stop when given a 200 response with an answer.

There is an unanswered question in this about whether there is a time
where we should stop polling as it has taken too long for an answer to
be generated, I expect we will handle this server side.

As this involves intervals, we make use of jasmine.clock as a way to
avoid the tests having to actually delay while we wait to make a call.