Specify pending answer url as a data attribute
If a user views a conversation and a question has been accepted but an
answer has not yet loaded we need a way to inform the application that
an answer needs to be loaded. The approach I've taken to do this, is to
set a data attribute on the chat-conversation module element (this was
chosen over putting it on an individual question element as that would
be harder to look-up in the dom and could imply multiple answers to
load).

I've used an ivar as a way to specify the data attribute properties
rather than the more common content_for, as content_for converts the
item to a string so is not suitable for specifying a hash.

In putting this together I can see that there are N+1 issues with the
queries. I'm going to look at addressing them in a separate PR.